move changelog to data.res
