+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/screens/about/ChangelogScreen.kt
@@ -26,7 +26,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import de.westnordost.streetcomplete.R
-import de.westnordost.streetcomplete.data.changelog.Changelog
+import de.westnordost.streetcomplete.data.res.Changelog
 import de.westnordost.streetcomplete.ui.common.BackIcon
 import de.westnordost.streetcomplete.ui.common.HtmlText
 import de.westnordost.streetcomplete.ui.ktx.plus
