+++ app/src/androidMain/kotlin/de/westnordost/streetcomplete/screens/main/messages/WhatsNewDialog.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import de.westnordost.streetcomplete.R
-import de.westnordost.streetcomplete.data.changelog.Changelog
+import de.westnordost.streetcomplete.data.res.Changelog
 import de.westnordost.streetcomplete.screens.about.ChangelogList
 import de.westnordost.streetcomplete.ui.common.dialogs.ScrollableAlertDialog
