+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/res/FlagAlignments.kt
@@ -2,6 +2,8 @@ package de.westnordost.streetcomplete.data.res
 
 import com.charleskorn.kaml.Yaml
 import de.westnordost.stretcomplete.resources.Res
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.decodeFromString
@@ -10,7 +12,7 @@ typealias FlagAlignments = Map<String, FlagAlignment>
 
 suspend fun Res.readFlagAlignments(): FlagAlignments {
     val yml = readBytes("files/flag_alignments.yml").decodeToString()
-    return Yaml.default.decodeFromString<FlagAlignments>(yml)
+    return withContext(Dispatchers.Default) { Yaml.default.decodeFromString<FlagAlignments>(yml) }
 }
 
 @Serializable
